# Optimise finding a content change for an email