The thinking here is that `first` will attempt to do some ordering based on the ID (which is a UUID, so the order has no meaning). `take` will simply go off and get the first one it finds.

I don't think we need to sort the `content_changes`, as we're excluding digests so they should all have approximately the same `created_at` (and we were sorting them by uuid before, so they've always been random).

A bit of explain analyze on staging suggests this might be a significant performance win (but I could be wrong - I'm not good with databases):

```sql
EXPLAIN ANALYZE
SELECT "content_changes".*
FROM "content_changes"
WHERE "content_changes"."id" IN
(
    SELECT "subscription_contents"."content_change_id"
    FROM "subscription_contents"
    WHERE "subscription_contents"."email_id" = 'REDACTED'
)
ORDER BY "content_changes"."id" ASC
LIMIT 1;
```

```
                                                                              QUERY PLAN
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=22.16..3393.12 rows=1 width=1314) (actual time=82.668..82.670 rows=1 loops=1)
   ->  Nested Loop Semi Join  (cost=22.16..2251820.27 rows=668 width=1314) (actual time=82.666..82.666 rows=1 loops=1)
         Join Filter: (content_changes.id = subscription_contents.content_change_id)
         Rows Removed by Join Filter: 661752
         ->  Index Scan using content_changes_pkey on content_changes  (cost=0.42..154618.22 rows=209036 width=1314) (actual time=0.005..4.430 rows=5211 loops=1)
         ->  Materialize  (cost=21.74..2663.01 rows=668 width=16) (actual time=0.000..0.006 rows=127 loops=5211)
               ->  Bitmap Heap Scan on subscription_contents  (cost=21.74..2659.67 rows=668 width=16) (actual time=0.022..0.043 rows=127 loops=1)
                     Recheck Cond: (email_id = '...'::uuid)
                     Heap Blocks: exact=4
                     ->  Bitmap Index Scan on index_subscription_contents_on_email_id  (cost=0.00..21.58 rows=668 width=0) (actual time=0.019..0.020 rows=127 loops=1)
                           Index Cond: (email_id = '...'::uuid)
 Planning time: 0.292 ms
 Execution time: 82.705 ms
(13 rows)
```

```sql
EXPLAIN ANALYZE
SELECT "content_changes".*
FROM "content_changes"
WHERE "content_changes"."id" IN
(
    SELECT "subscription_contents"."content_change_id"
    FROM "subscription_contents"
    WHERE "subscription_contents"."email_id" = 'REDACTED'
)
-- NOTE: order by is removed
LIMIT 1;
```

```
                                                                              QUERY PLAN
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=2661.76..2667.95 rows=1 width=1314) (actual time=0.074..0.075 rows=1 loops=1)
   ->  Nested Loop  (cost=2661.76..6801.08 rows=668 width=1314) (actual time=0.073..0.073 rows=1 loops=1)
         ->  HashAggregate  (cost=2661.34..2666.93 rows=559 width=16) (actual time=0.066..0.066 rows=1 loops=1)
               Group Key: subscription_contents.content_change_id
               ->  Bitmap Heap Scan on subscription_contents  (cost=21.74..2659.67 rows=668 width=16) (actual time=0.024..0.039 rows=127 loops=1)
                     Recheck Cond: (email_id = '...'::uuid)
                     Heap Blocks: exact=4
                     ->  Bitmap Index Scan on index_subscription_contents_on_email_id  (cost=0.00..21.58 rows=668 width=0) (actual time=0.020..0.020 rows=127 loops=1)
                           Index Cond: (email_id = '...'::uuid)
         ->  Index Scan using content_changes_pkey on content_changes  (cost=0.42..7.39 rows=1 width=1314) (actual time=0.005..0.005 rows=1 loops=1)
               Index Cond: (id = subscription_contents.content_change_id)
 Planning time: 0.277 ms
 Execution time: 0.127 ms
(13 rows)
```