@kevindew 2020-06-10
Nice work 🚀  that is an amazing reduction in time. TIL on the `take` method too. Looks super handy for this project.

It's bananas how much of a difference the `ORDER BY` seems to make given we never anticipate it to ever return more than one record so would never need any ordering. 

Since the query plan is still a bit grim in what it expects to do, do we think this performance might fluctuate somewhat?

The approaches Chris and I looked at involved leveraging the [`SubscriptionContent.where(email: email).digest.exists?`](https://github.com/alphagov/email-alert-api/pull/1271/files#diff-0f17b24554bd2aa9302470840a10845eR65) query to avoid the need for a join altogether and switch the created_at pull to be a primary key lookup.

Roughly it comes out something like:

```
      content_change_id = SubscriptionContent.where(email: email).immediate.pick(:content_change_id)
      return unless content_change_id

      created_at = ContentChange.where(id: content_change_id).pick(:created_at)
      return unless created_at

      difference = (time - created_at) * 1000
      namespace = "content_change_created_to_first_delivery_attempt"
      timing(namespace, difference)
```

Plan wise the two queries come out at:

```
EXPLAIN ANALYZE SELECT content_change_id FROM subscription_contents WHERE subscription_contents.email_id = 'cbd0acc0-fb98-424d-8415-bb50ba5e117f' AND "subscription_contents"."digest_run_subscriber_id" IS NULL LIMIT 1;

---------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.57..9.39 rows=1 width=16) (actual time=0.012..0.012 rows=1 loops=1)
   ->  Index Scan using index_subscription_contents_on_email_id on subscription_contents  (cost=0.57..2507.16 rows=284 width=16) (actual time=0.012..0.013 rows=1 loops=1)
         Index Cond: (email_id = 'cbd0acc0-fb98-424d-8415-bb50ba5e117f'::uuid)
         Filter: (digest_run_subscriber_id IS NULL)
 Planning time: 0.059 ms
 Execution time: 0.022 ms
(6 rows)
```

and

```
EXPLAIN ANALYZE SELECT "content_changes"."created_at" FROM "content_changes" WHERE "content_changes"."id" = '8521d24b-79ec-4d83-b2ad-62f3ee013a02' LIMIT 1;
--------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.42..8.44 rows=1 width=8) (actual time=0.015..0.015 rows=1 loops=1)
   ->  Index Scan using content_changes_pkey on content_changes  (cost=0.42..8.44 rows=1 width=8) (actual time=0.013..0.013 rows=1 loops=1)
         Index Cond: (id = '8521d24b-79ec-4d83-b2ad-62f3ee013a02'::uuid)
 Planning time: 0.055 ms
 Execution time: 0.026 ms
(5 rows)
```

---

Tangentially I feel like at some point in this week we should move this query lookup out of the metrics service (if we need it, I know Chris is investigating removing need for a query altogether). It seems rather surprising that a call to a `MetricsService` talks to the db.