@richardTowers 2020-06-10
Yeah, that approach looks sensible to me too - probably a bit easier to understand than this monstrosity.

I did a bit of research around the `first` / `take` difference - the ordering on first is useful when you're using auto-increment ids, as the ordering will get you the "first" entry that was inserted.  It's pointless if you're using `uuids` though. I think that might be a good general lesson for us, as we use uuids a lot.

Strongly agree with moving the query outside of the metric service too - it's super surprising for a metric write to be messing with the DB, which I think is a big part of why we didn't spot this performance issue earlier.